+++ crates/tuirealm/src/lib.rs
@@ -91,3 +91,8 @@ pub use self::core::props::{self, AttrValue, Attribute, Props};
 pub use self::core::subscription::{EventClause as SubEventClause, Sub, SubClause};
 pub use self::core::{Component, MockComponent, State, StateValue, Update, ViewError, command};
 pub use self::ratatui::Frame;
+
+// export async trait for async-ports
+#[cfg(feature = "async-ports")]
+#[cfg_attr(docsrs, doc(cfg(feature = "async-ports")))]
+pub use async_trait::async_trait;