+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [2.3.0](#230)
   - [2.2.0](#220)
   - [2.1.0](#210)
   - [2.0.3](#203)
@@ -40,6 +41,40 @@
   - [0.1.0](#010)
 
 ---
+
+## 2.3.0
+
+Released on 17/05/2025
+
+- [Issue 94](https://github.com/veeso/tui-realm/issues/94): Added support for **Async Ports**. It is now possible to create async ports by implementing the `PollAsync` trait and then by passing the `tokio::rt::Handle` to the `EventListenerCfg::port` method.
+
+    ```rust
+    let handle = Arc::new(tokio::rt::Handle::current());
+
+    let event_listener = EventListenerCfg::default()
+        .crossterm_input_listener(Duration::from_millis(10), 3)
+        .add_async_port(
+            Box::new(AsyncPort::new()),
+            Duration::from_millis(1000),
+            1,
+            &handle,
+        );
+
+    #[tuirealm::async_trait]
+    impl PollAsync<UserEvent> for AsyncPort {
+        async fn poll(&self) -> ListenerResult<Option<Event<UserEvent>>> {
+            let result = self.write_file().await;
+
+            Ok(Some(Event::User(UserEvent::WroteFile(result))))
+        }
+    }
+    ```
+
+    ⚠️ `PollAsync` takes a `&self` reference, and not a `&mut self` reference as `Poll` does; this is because it is not possible to have mutable references with async in this case. Also consider that `UserEvent` must be `Send` and `Sync` if you want to use it in an async context.
+
+    ‼️ **PollAsync** is featured behind the `async-ports` feature.
+
+    ❗ You can find an example of this at `examples/async_ports.rs`.
 
 ## 2.2.0